+++ source/main/menus/win-linux.ts
@@ -7,12 +7,10 @@ import { environment } from '../environment';
 import { NOTIFICATIONS } from '../../common/ipc/constants';
 import { showUiPartChannel } from '../ipc/control-ui-parts';
 import { generateSupportRequestLink } from '../../common/utils/reporting';
-import { getRtsFlags } from '../utils/rtsFlags';
+import { buildKnownIssueFixesSubmenu } from './submenuBuilders';
+import { WalletSettingsStateEnum } from '../../common/ipc/api';
 
 const id = 'menu';
-const { isWindows, isBlankScreenFixActive, network } = environment;
-const rtsFlags = getRtsFlags(network);
-const rtsFlagsEnabled: boolean = !!rtsFlags?.length && rtsFlags.length > 0;
 export const winLinuxMenu = (
   app: App,
   window: BrowserWindow,
@@ -20,6 +18,7 @@ export const winLinuxMenu = (
   translations: {},
   locale: string,
   isNavigationEnabled: boolean,
+  walletSettingsState: WalletSettingsStateEnum,
   translation: (...args: Array<any>) => any = getTranslation(translations, id)
 ) => [
   {
@@ -130,12 +129,15 @@ export const winLinuxMenu = (
           actions.openWalletSettingsPage();
         },
 
-        enabled: isNavigationEnabled,
+        enabled:
+          isNavigationEnabled &&
+          walletSettingsState === WalletSettingsStateEnum.enabled,
+        visible: walletSettingsState !== WalletSettingsStateEnum.hidden,
       },
       {
         type: 'separator',
       },
-      isWindows
+      environment.isWindows
         ? {
             label: translation('view.toggleFullScreen'),
             accelerator: 'F11',
@@ -170,33 +172,7 @@ export const winLinuxMenu = (
   {
     label: translation('helpSupport'),
     submenu: compact([
-      {
-        label: translation('helpSupport.knownIssues'),
-
-        click() {
-          const faqLink = translation('helpSupport.knownIssuesUrl');
-          shell.openExternal(faqLink);
-        },
-      },
-      {
-        label: translation('helpSupport.blankScreenFix'),
-        type: 'checkbox',
-        checked: isBlankScreenFixActive,
-
-        click(item) {
-          actions.toggleBlankScreenFix(item);
-        },
-      },
-      {
-        label: translation('helpSupport.usingRtsFlags'),
-        type: 'checkbox',
-        checked: rtsFlagsEnabled,
-
-        click(item) {
-          actions.setRtsFlags(!rtsFlagsEnabled);
-          item.checked = rtsFlagsEnabled;
-        },
-      },
+      ...buildKnownIssueFixesSubmenu(actions, translations, translation),
       {
         type: 'separator',
       },