+++ source/renderer/app/ipc/getBlockSyncChannel.ts
@@ -1,0 +1,12 @@
+import { GET_BLOCK_SYNC_STATUS_CHANNEL } from '../../../common/ipc/api';
+import type {
+  GetBlockSyncProgressRendererRequest,
+  GetBlockSyncProgressMainResponse,
+} from '../../../common/ipc/api';
+import { RendererIpcChannel } from './lib/RendererIpcChannel';
+
+export const getBlockSyncProgressChannel: // IpcChannel<Incoming, Outgoing>
+RendererIpcChannel<
+  GetBlockSyncProgressMainResponse,
+  GetBlockSyncProgressRendererRequest
+> = new RendererIpcChannel(GET_BLOCK_SYNC_STATUS_CHANNEL);